feat: 提供一个快速创建\覆盖fetcher的方法
