+++ packages/amis-core/src/utils/index.ts
@@ -54,6 +54,7 @@ export * from './toNumber';
 export * from './decodeEntity';
 export * from './style-helper';
 export * from './resolveCondition';
+export * from './buildFetcher';
 
 import animation from './Animation';
